requirements.txt
scipy == 1.5.4
numpy == 1.19.4 
matplotlib == 3.3.3
lmfit == 1.0.1
pyyaml